Enable browser Arduino backend, frictionless dev, and browser caching
Make the Arduino backend work the same in desktop and browser, and remove the
setup friction that left a packaged build with 'no Arduino service':

- Extract the proven WebSocket Arduino logic into WebSocketArduinoService; both
  ElectronArduinoService and WebArduinoService now extend it. The browser build
  can compile/upload/serial by connecting to a local tinyService on
  ws://localhost:3000 (override via localStorage['tinyservice.url']).
- ServiceManager: in dev, use the arduino-cli fetched into vendor/ (falling back
  to PATH) so a fresh clone works without a global install.
- fetch-arduino-cli.mjs: add 'current' to fetch only the host platform; wire a
  predev hook so 'npm run dev' auto-provisions it.
- App: surface a clear 'Arduino backend not running' toast instead of failing
  silently; reconnect is automatic.
- Browser durability: webCache (IndexedDB) mirrors every web save and serves as
  a read fallback when File System Access permissions are lost on reload.
- Docs: README + INTEGRATION_GUIDE now describe public-npm install (no token)
  and the desktop-vs-browser backend model.
